+++ packages/dma-library/src/types/index.ts
@@ -130,7 +130,9 @@ export type {
   Erc4626StrategyAddresses,
   Erc4626WithdrawPayload,
   Erc4626WithdrawStrategy,
+  IErc4626Position,
 } from './common'
+export { Erc4626Position, FeeType } from './common'
 export type {
   EarnCumulativesData,
   EarnCumulativesRawData,
