+++ packages/dma-library/src/views/common/index.ts
@@ -1,3 +1,2 @@
 export { getErc4626Position } from './erc4626'
 export * from './get-buying-power'
-export { Erc4626Position } from './types'
